+++ kinto_emailer/tests/test_includeme.py
@@ -7,7 +7,7 @@ import configparser
 
 from kinto import main as kinto_main
 from kinto.core.events import AfterResourceChanged
-from kinto.core.testing import BaseWebTest
+from kinto.core.testing import BaseWebTest, get_user_headers
 from kinto_emailer import get_messages, send_notification
 
 
@@ -34,7 +34,7 @@ COLLECTION_RECORD = {
 }
 
 
-class PluginSetupTest(BaseWebTest, unittest.TestCase):
+class EmailerTest(BaseWebTest, unittest.TestCase):
     entry_point = kinto_main
     api_prefix = "v1"
     config = 'config/kinto.ini'
@@ -48,6 +48,8 @@ class PluginSetupTest(BaseWebTest, unittest.TestCase):
             settings.update(extras)
         return settings
 
+
+class PluginSetupTest(EmailerTest):
     def test_capability_is_exposed(self):
         resp = self.app.get('/')
         capabilities = resp.json['capabilities']
@@ -182,11 +184,41 @@ class GetMessagesTest(unittest.TestCase):
         messages = get_messages(collection_record, payload)
         assert len(messages) == 1
 
+    def test_get_messages_can_filter_by_event_class(self):
+        collection_record = {
+            'kinto-emailer': {
+                'hooks': [{
+                    'event': 'kinto_signer.events.ReviewRequested',
+                    'template': 'Poll changed.',
+                    'recipients': ['me@you.com'],
+                }]
+            }
+        }
+        payload = {
+            'event': 'kinto.core.events.AfterResourceChanged',
+            'resource_name': 'record',
+            'action': 'create'
+        }
+        messages = get_messages(collection_record, payload)
+        assert len(messages) == 0
+        payload = {
+            'event': 'kinto_signer.events.ReviewRequested',
+            'resource_name': 'record',
+            'action': 'create'
+        }
+        messages = get_messages(collection_record, payload)
+        assert len(messages) == 1
 
 
 class SendNotificationTest(unittest.TestCase):
     def test_send_notification_does_not_call_the_mailer_if_no_message(self):
         event = mock.MagicMock()
+        event.payload = {
+            'resource_name': 'record',
+            'action': 'update',
+            'bucket_id': 'default',
+            'collection_id': 'foobar'
+        }
         event.request.registry.storage.get.return_value = {}
 
         with mock.patch('kinto_emailer.get_mailer') as get_mailer:
@@ -205,4 +237,59 @@ class SendNotificationTest(unittest.TestCase):
 
         with mock.patch('kinto_emailer.get_mailer') as get_mailer:
             send_notification(event)
+            assert get_mailer().send.called
+
+
+class SignerEventsTest(EmailerTest):
+    def get_app_settings(self, extras=None):
+        settings = super(SignerEventsTest, self).get_app_settings(extras)
+        settings['kinto.includes'] += ' kinto_signer'
+        settings['kinto.signer.resources'] = (
+            '/buckets/staging/collections/addons;'
+            '/buckets/blocklists/collections/addons')
+        settings['kinto.signer.group_check_enabled'] = 'false'
+        settings['kinto.signer.to_review_enabled'] = 'true'
+        settings['kinto.signer.signer_backend'] = 'kinto_signer.signer.autograph'
+        settings['kinto.signer.autograph.server_url'] = 'http://localhost:8000'
+        settings['kinto.signer.autograph.hawk_id'] = 'not-used-because-mocked'
+        settings['kinto.signer.autograph.hawk_secret'] = 'not-used-because-mocked'
+        return settings
+
+    def setUp(self):
+        collection = {
+            'kinto-emailer': {
+                'hooks': [{
+                    'event': 'kinto_signer.events.ReviewRequested',
+                    'template': '{user_id} requested review on {uri}.',
+                    'recipients': ['me@you.com'],
+                }]
+            }
+        }
+        self.headers = dict(self.headers, **get_user_headers('nous'))
+        self.app.put('/buckets/staging', headers=self.headers)
+        self.app.put_json('/buckets/staging/collections/addons',
+                          {'data': collection},
+                          headers=self.headers)
+        self.app.post_json('/buckets/staging/collections/addons/records',
+                           {'data': {'age': 42}},
+                           headers=self.headers)
+        self._patch_autograph()
+
+    def _patch_autograph(self):
+        # Patch calls to Autograph.
+        patch = mock.patch('kinto_signer.signer.autograph.requests')
+        mocked = patch.start()
+        mocked.post.return_value.json.return_value = [{
+            "signature": "",
+            "hash_algorithm": "",
+            "signature_encoding": "",
+            "content-signature": "",
+            "x5u": ""}]
+        return patch
+
+    def test_email_is_when_review_is_requested(self):
+        with mock.patch('kinto_emailer.get_mailer') as get_mailer:
+            self.app.patch_json('/buckets/staging/collections/addons',
+                                {'data': {'status': 'to-review'}},
+                                headers=self.headers)
             assert get_mailer().send.called