+++ CHANGELOG.rst
@@ -9,3 +9,4 @@ This document describes changes between each past release.
 **Initial version**
 
 - Use a list of hooks to configure emails bound to notifications (fixes #11)
+- Support *kinto-signer* events (fixes #14)